requirements.txt
# Core LLM and Agent Framework (v1.0+)
langchain>=1.0.0
langchain-community>=0.4.0
langchain-core>=1.0.0
langchain-openai>=1.0.0

# MCP Protocol for connecting to MCP servers
langchain-mcp-adapters>=0.1.12

# LangGraph Core and Server (v1.0+)
langgraph>=1.0.0
langgraph-cli[inmem]>=0.1.0
langgraph-checkpoint-postgres>=2.0.0

# PostgreSQL driver with async support and connection pooling
psycopg[binary,pool]>=3.0.0

# Additional dependencies
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
